--- app/src/main/java/com/example/memorygame/view/contents/Contents.kt
@@ -1,7 +1,0 @@
-package com.example.memorygame.view.contents
-
-object Contents {
-    const val GAME_READY_MODE = 0
-    const val GAME_START_MODE = 1
-    const val GAME_ROUND_TIME = 30000L
-}
