--- app/src/main/res/drawable/info_background.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android">
-    <solid android:color="#60BDBCBC" />
-<corners android:radius="8dp"/>
-
-</shape>
