+++ app/src/main/res/layout/fragment_main.xml
@@ -1,91 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
-
-<layout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
-<androidx.constraintlayout.widget.ConstraintLayout
-    android:id="@+id/main"
+<layout xmlns:android="http://schemas.android.com/apk/res/android">
+  
+  <androidx.constraintlayout.widget.ConstraintLayout
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".view.dest.MainFragment"
-    android:background="#D0000000">
-
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/recyclerView"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="100dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
-        app:spanCount="3"
-        tools:listitem="@layout/viewholder"/>
-
-    <LinearLayout
-        android:id="@+id/linearLayout"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="32dp"
-        android:background="@drawable/info_background"
-        android:gravity="center_vertical"
-        android:orientation="horizontal"
-        app:layout_constraintBottom_toTopOf="@+id/recyclerView"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent">
-
-        <TextView
-            android:id="@+id/textView_info"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_margin="10dp"
-            android:text= "@string/tv_gameReady"
-            android:textColor="#FFF"
-            android:textSize="54sp"
-            android:textStyle="bold" />
-
-        <ImageView
-            android:id="@+id/imageView_imageCheck"
-            android:layout_width="54dp"
-            android:layout_height="54dp"
-            android:layout_marginEnd="10dp"
-            android:visibility="gone"
-            tools:src="@drawable/ic_launcher_background" />
-    </LinearLayout>
-
-    <View
-        android:id="@+id/view_timeBar"
-        android:layout_width="0dp"
-        android:layout_height="5dp"
-        android:background="#fff"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="@id/view_distanceMeasurement"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <View
-        android:id="@+id/view_distanceMeasurement"
-        android:layout_width="1dp"
-        android:layout_height="0dp"
-        app:layout_constraintEnd_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <TextView
-        android:id="@+id/textView_gameRound"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="24dp"
-        android:background="@drawable/info_background"
-        android:paddingStart="15dp"
-        android:paddingEnd="15dp"
-        android:textColor="#fff"
-        android:textSize="30sp"
-        android:textStyle="bold"
-        app:layout_constraintBottom_toTopOf="@+id/linearLayout"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/view_timeBar"
-        tools:text="0/3" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
+    android:layout_height="match_parent">
+  
+  </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
