--- app/src/main/res/navigation/navi.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto" android:id="@+id/navi"
-    app:startDestination="@id/mainFragment">
-
-    <fragment
-        android:id="@+id/mainFragment"
-        android:name="com.example.i4_memorygame.view.dest.MainFragment"
-        android:label="MainFragment" >
-        <argument
-            android:name="layoutWidth"
-            app:argType="integer" />
-    </fragment>
-</navigation>
