--- app/src/main/res/values-ko-rKR/strings.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-  <string name="app_name">메모리게임</string>
-</resources>
