+++ app/src/main/res/values/strings.xml
@@ -1,9 +1,3 @@
 <resources>
-    <string name="app_name">14_MemoryGame</string>
-    <string name="tv_gameReady">준비</string>
-    <string name="tv_remember">기억하세요</string>
-    <string name="tv_find">찾기:</string>
-    <string name="tv_answer">정답입니다!</string>
-    <string name="tv_wrongAnswer">틀렸습니다..</string>
-    <string name="tv_currentPage">(%d)/3</string>
+  <string name="app_name">MemoryGame</string>
 </resources>
